+++ src/lib/index.ts
@@ -4,5 +4,6 @@ import BlogPage, { type Config } from "./components/BlogPage.svelte";
 import Image from "./components/post/Image.svelte";
 import Audio from "./components/post/Audio.svelte";
 import FileAttachment from "./components/post/FileAttachment.svelte";
+import Quote from "./components/post/Quote.svelte";
 
-export {BlogPage, type Config, Image, Audio, FileAttachment};
+export {BlogPage, type Config, Image, Audio, FileAttachment, Quote};