+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The header title can now have an URL to point to.
 - The post metadata div now have an exposed class.
+- Added Quote post component.
 
 ### Fixed
 
@@ -18,7 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### TODO
 
-- quote component
 - note component
 
 ## [0.0.2] - 2024-10-26